Move `LocalFeedImageDataLoader` and `FeedimageDataStore` to production
